Add correct LGPL 2.1 license file and add SPDX boilerplates
Rename the license file to COPYING.LIB which is the common convention
for the LGPL.  Place the latest text copy of the LGPL 2.1 license in
COPYING.LIB.  Add a symlink for 'LICENSE' pointing to COPYING.LIB
since the project previously used the LICENSE file to hold the
license.

Add copyright and SPDX license identifier boilerplate comments to the
src and test files.

Fixes: #45